backend/model/gemini.py
import os
from google import genai
from dotenv import load_dotenv
load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

response = client.models.generate_content(
    model="gemini-2.5-flash",
    contents="give info about DISEASE PREVENTION AND CONTROL SUMMIT AMERICA 2025, what is it?",
)

print(response.text)