backend/requirements.txt
beautifulsoup4
python-dotenv
requests
lxml
pymongo
pydantic
fastapi
uvicorn
google-genai
groq
tavily
anthropic
langchain_anthropic 